Create Netflix Titles and Info
This file contains a list of queries I used to explore data from Netflix's database of movies and TV shows. I calculated the number of movies in Netflix's database, as well as find the most recently added batch of movies/TV shows. You can also see that I found the oldest movie added in the database, and was able to find the names of titles and their directors.